fix: Resolve dtype mismatch in GistLlama.generate()
**Error**: RuntimeError: Index put requires the source and destination dtypes
match, got Float for the destination and BFloat16 for the source.

**Root Cause**: In generate() method, gist_embedding parameter is in bfloat16
but inputs_embeds from embed_tokens() can be in float32, causing dtype mismatch
when replacing gist tokens.

**Fix**: Convert gist_embedding to match inputs_embeds dtype before assignment:
`inputs_embeds[gist_mask] = self.gist_embedding.to(inputs_embeds.dtype)`

This ensures dtypes match regardless of model precision (fp32, fp16, bfloat16).

+++ compressions/train_gist_faithful.py
@@ -265,7 +265,8 @@ class GistLlama(nn.Module):
         inputs_embeds = self.model.model.embed_tokens(input_ids)
         gist_mask = (input_ids == self.gist_token_id)
         if gist_mask.any():
-            inputs_embeds[gist_mask] = self.gist_embedding
+            # Ensure dtype matches (gist_embedding might be bfloat16 while inputs_embeds is float32)
+            inputs_embeds[gist_mask] = self.gist_embedding.to(inputs_embeds.dtype)
 
         # Apply gist attention mask if provided
         if attention_mask_gist is not None:
